+++ .gitignore
@@ -1,3 +1,3 @@
 node_modules
 package-lock.json
-src/components/tests/coverage
+src/tests/coverage
